LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-referrer.any-expected.txt

CONSOLE MESSAGE: line 32: ReferenceError: Can't find variable: get_host_info

Harness Error (FAIL), message = ReferenceError: Can't find variable: get_host_info
